+++ src/admin/dto/update-pay-policy.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/swagger';
+import { CreatePayPolicyDto } from './create-pay-policy.dto';
+
+export class UpdatePayPolicyDto extends PartialType(CreatePayPolicyDto) {}